+++ src/main/java/org/cesarb/vowels/Pair.java
@@ -1,4 +1,0 @@
-package org.cesarb.vowels;
-
-public class Pair {
-}