+++ src/main/java/org/cesarb/vowels/Triple.java
@@ -1,0 +1,25 @@
+package org.cesarb.vowels;
+
+public class Triple<A, B, C> {
+    private A first;
+    private B second;
+    private C third;
+
+    public Triple(A first, B second, C third) {
+        this.first = first;
+        this.second = second;
+        this.third = third;
+    }
+
+    public A first() {
+        return first;
+    }
+
+    public B second() {
+        return second;
+    }
+
+    public C third() {
+        return third;
+    }
+}